remove unnecessary helpers and fix the table index EOM
